Move InternalBytecode.h to InternalJavaScript
Summary: Move InternalBytecode.h to InternalJavaScript directory.

Reviewed By: avp

Differential Revision: D53023942

fbshipit-source-id: c084cf56f0278d9ea9e56fe9570f870372e38941
