+++ lib/InternalJavaScript/01-Promise.js
@@ -524,8 +524,8 @@
   };
 
   // @nolint
-  // This file is used to generate InternalBytecode/Promise.js
-  // See InternalBytecode/README.md for more details.
+  // This file is used to generate InternalJavaScript/Promise.js
+  // See InternalJavaScript/README.md for more details.
